💄 스캔 목록 UI 수정.
